[Ray] Optimize Ray CI execution time and stability (#3102)
(cherry picked from commit b685973b92d0cbc2e9746406135b0ad17fe1d727)
