+++ mars/deploy/oscar/tests/test_ray.py
@@ -16,30 +16,16 @@ import asyncio
 import copy
 import operator
 import os
-import subprocess
-import sys
-import tempfile
-import threading
-import time
 from functools import reduce
 
 import numpy as np
-import pandas as pd
 import pytest
 
 import mars
-from .... import oscar as mo
 from .... import tensor as mt
 from .... import dataframe as md
-from ....oscar.backends.ray.utils import (
-    process_address_to_placement,
-    process_placement_to_address,
-    kill_and_wait,
-)
 from ....oscar.errors import ReconstructWorkerError
 from ....serialization.ray import register_ray_serializers
-from ....services.cluster import ClusterAPI
-from ....services.scheduling.supervisor.autoscale import AutoscalerActor
 from ....tests.core import require_ray, mock, DICT_NOT_EMPTY
 from ....utils import lazy_import
 from ..ray import (
@@ -98,7 +84,6 @@ async def create_cluster(request):
     ray_config = _load_config(CONFIG_FILE)
     ray_config.update(param.get("config", {}))
     client = await new_cluster(
-        "test_cluster",
         supervisor_mem=1 * 1024**3,
         worker_num=2,
         worker_cpu=2,
@@ -135,25 +120,25 @@ async def create_cluster(request):
     ],
 )
 @pytest.mark.asyncio
-async def test_execute(ray_start_regular, create_cluster, config):
+async def test_execute(ray_start_regular_shared, create_cluster, config):
     await test_local.test_execute(create_cluster, config)
 
 
 @require_ray
 @pytest.mark.asyncio
-async def test_iterative_tiling(ray_start_regular, create_cluster):
+async def test_iterative_tiling(ray_start_regular_shared, create_cluster):
     await test_local.test_iterative_tiling(create_cluster)
 
 
 @require_ray
 @pytest.mark.asyncio
-async def test_execute_describe(ray_start_regular, create_cluster):
+async def test_execute_describe(ray_start_regular_shared, create_cluster):
     await test_local.test_execute_describe(create_cluster)
 
 
 @require_ray
 @pytest.mark.asyncio
-async def test_fetch_infos(ray_start_regular, create_cluster):
+async def test_fetch_infos(ray_start_regular_shared, create_cluster):
     await test_local.test_fetch_infos(create_cluster)
     df = md.DataFrame(mt.random.RandomState(0).rand(5000, 1, chunk_size=1000))
     df.execute()
@@ -165,7 +150,7 @@ async def test_fetch_infos(ray_start_regular, create_cluster):
 
 @require_ray
 @pytest.mark.asyncio
-def test_sync_execute(ray_start_regular, create_cluster):
+def test_sync_execute(ray_start_regular_shared, create_cluster):
     client = create_cluster[0]
     assert client.session
     session = new_session(address=client.address)
@@ -247,59 +232,6 @@ def new_ray_session_test():
 
 
 @require_ray
-def test_ray_client(ray_start_regular):
-    server_code = """import time
-import ray.util.client.server.server as ray_client_server
-
-server = ray_client_server.init_and_serve("{address}", num_cpus=20)
-print("OK", flush=True)
-while True:
-    time.sleep(1)
-"""
-
-    address = "127.0.0.1:50051"
-
-    with tempfile.NamedTemporaryFile(mode="w", suffix=".py") as f:
-        f.write(server_code.format(address=address))
-        f.flush()
-
-        proc = subprocess.Popen([sys.executable, "-u", f.name], stdout=subprocess.PIPE)
-
-        try:
-
-            def _check_ready(expect_exit=False):
-                while True:
-                    line = proc.stdout.readline()
-                    if proc.returncode is not None:
-                        if expect_exit:
-                            break
-                        raise Exception(
-                            f"Failed to start ray server at {address}, "
-                            f"the return code is {proc.returncode}."
-                        )
-                    if b"OK" in line:
-                        break
-
-            # Avoid ray.init timeout.
-            _check_ready()
-
-            # Avoid blocking the subprocess when the stdout pipe is full.
-            t = threading.Thread(target=_check_ready, args=(True,))
-            t.start()
-
-            ray.init(f"ray://{address}")
-            ray._inside_client_test = True
-            try:
-                new_ray_session_test()
-            finally:
-                ray._inside_client_test = False
-                ray.shutdown()
-        finally:
-            proc.kill()
-            proc.wait()
-
-
-@require_ray
 @pytest.mark.parametrize(
     "test_option",
     [
@@ -310,7 +242,7 @@ while True:
     ],
 )
 @pytest.mark.asyncio
-async def test_optional_supervisor_node(ray_start_regular, test_option):
+async def test_optional_supervisor_node(ray_start_regular_shared, test_option):
     import logging
 
     logging.basicConfig(level=logging.INFO)
@@ -357,7 +289,7 @@ async def test_optional_supervisor_node(ray_start_regular, test_option):
     ],
 )
 @pytest.mark.asyncio
-async def test_web_session(ray_start_regular, create_cluster, config):
+async def test_web_session(ray_start_regular_shared, create_cluster, config):
     client = create_cluster[0]
     await test_local.test_web_session(create_cluster, config)
     web_address = client.web_address
@@ -381,15 +313,15 @@ async def test_web_session(ray_start_regular, create_cluster, config):
     ],
 )
 @pytest.mark.asyncio
-async def test_load_third_party_modules(ray_start_regular, config_exception):
+async def test_load_third_party_modules(ray_start_regular_shared, config_exception):
     third_party_modules_config, expected_exception = config_exception
     config = _load_config()
 
     config["third_party_modules"] = third_party_modules_config
     with expected_exception:
         await new_cluster(
-            worker_num=2,
-            worker_cpu=2,
+            worker_num=1,
+            worker_cpu=1,
             worker_mem=1 * 1024**3,
             config=config,
         )
@@ -410,7 +342,7 @@ async def test_load_third_party_modules(ray_start_regular, config_exception):
     indirect=True,
 )
 @pytest.mark.asyncio
-def test_load_third_party_modules2(ray_start_regular, create_cluster):
+def test_load_third_party_modules2(ray_start_regular_shared, create_cluster):
     client = create_cluster[0]
     assert client.session
     session = new_session(address=client.address)
@@ -429,18 +361,18 @@ def test_load_third_party_modules2(ray_start_regular, create_cluster):
 @require_ray
 @pytest.mark.asyncio
 async def test_load_third_party_modules_from_config(
-    ray_start_regular, cleanup_third_party_modules_output  # noqa: F811
+    ray_start_regular_shared, cleanup_third_party_modules_output  # noqa: F811
 ):
     client = await new_cluster(
         supervisor_mem=1 * 1024**3,
-        worker_num=2,
-        worker_cpu=2,
+        worker_num=1,
+        worker_cpu=1,
         worker_mem=1 * 1024**3,
         config=CONFIG_THIRD_PARTY_MODULES_TEST_FILE,
     )
     async with client:
-        # 1 supervisor, 2 worker main pools, 4 worker sub pools.
-        assert len(get_output_filenames()) == 7
+        # 1 supervisor, 1 worker main pools, 1 worker sub pools.
+        assert len(get_output_filenames()) == 3
 
 
 @require_ray
@@ -460,84 +392,6 @@ def test_load_config():
     with pytest.raises(ValueError):
         _load_config({"scheduling.autoscale.enabled": True, "scheduling.autoscale": {}})
     assert _load_config(CONFIG_FILE)["session"]["custom_log_dir"] == "auto"
-
-
-@pytest.mark.parametrize(
-    "ray_large_cluster", [{"num_nodes": 1, "num_cpus": 3}], indirect=True
-)
-@require_ray
-@pytest.mark.asyncio
-async def test_request_worker(ray_large_cluster):
-    worker_cpu, worker_mem = 1, 100 * 1024**2
-    client = await new_cluster(
-        worker_num=0, worker_cpu=worker_cpu, worker_mem=worker_mem
-    )
-    async with client:
-        cluster_state_ref = client._cluster._cluster_backend.get_cluster_state_ref()
-        # Note that supervisor took one node
-        workers = await asyncio.gather(
-            *[cluster_state_ref.request_worker(timeout=5) for _ in range(2)]
-        )
-        assert all(worker is not None for worker in workers)
-        assert not await cluster_state_ref.request_worker(timeout=5)
-        release_workers = [
-            cluster_state_ref.release_worker(worker) for worker in workers
-        ]
-        # Duplicate release workers requests should be handled.
-        release_workers.extend(
-            [cluster_state_ref.release_worker(worker) for worker in workers]
-        )
-        await asyncio.gather(*release_workers)
-        assert await cluster_state_ref.request_worker(timeout=5)
-        cluster_state_ref.reconstruct_worker()
-
-
-@pytest.mark.parametrize(
-    "ray_large_cluster", [{"num_nodes": 1, "num_cpus": 3}], indirect=True
-)
-@require_ray
-@pytest.mark.asyncio
-async def test_reconstruct_worker(ray_large_cluster):
-    worker_cpu, worker_mem = 1, 100 * 1024**2
-    client = await new_cluster(
-        worker_num=0, worker_cpu=worker_cpu, worker_mem=worker_mem
-    )
-    async with client:
-        cluster_api = await ClusterAPI.create(client._cluster.supervisor_address)
-        worker = await cluster_api.request_worker(timeout=5)
-        pg_name, bundle_index, process_index = process_address_to_placement(worker)
-        worker_sub_pool = process_placement_to_address(
-            pg_name, bundle_index, process_index + 1
-        )
-
-        worker_actor = ray.get_actor(worker)
-        worker_pid = await worker_actor.getpid.remote()
-        # the worker pool actor should be destroyed even we get actor.
-        worker_sub_pool_actor = ray.get_actor(worker_sub_pool)
-        worker_sub_pool_pid = await worker_sub_pool_actor.getpid.remote()
-
-        # kill worker main pool
-        await kill_and_wait(ray.get_actor(worker))
-
-        # duplicated reconstruct worker request can be handled.
-        await asyncio.gather(
-            cluster_api.reconstruct_worker(worker),
-            cluster_api.reconstruct_worker(worker),
-        )
-        worker_actor = ray.get_actor(worker)
-        new_worker_pid = await worker_actor.getpid.remote()
-        worker_sub_pool_actor = ray.get_actor(worker_sub_pool)
-        new_worker_sub_pool_pid = await worker_sub_pool_actor.getpid.remote()
-        assert new_worker_pid != worker_pid
-        assert new_worker_sub_pool_pid != worker_sub_pool_pid
-
-        # the compute should be ok after the worker is reconstructed.
-        raw = np.random.RandomState(0).rand(10, 5)
-        a = mt.tensor(raw, chunk_size=5).sum(axis=1)
-        b = a.execute(show_progress=False)
-        assert b is a
-        result = a.fetch()
-        np.testing.assert_array_equal(result, raw.sum(axis=1))
 
 
 @require_ray
@@ -595,138 +449,9 @@ async def test_release_worker_during_reconstructing_worker(
     release_task.cancel()
 
 
-@pytest.mark.parametrize(
-    "ray_large_cluster", [{"num_nodes": 2, "num_cpus": 4}], indirect=True
-)
-@pytest.mark.parametrize("init_workers", [0, 1])
 @require_ray
 @pytest.mark.asyncio
-async def test_auto_scale_out(ray_large_cluster, init_workers: int):
-    client = await new_cluster(
-        worker_num=init_workers,
-        worker_cpu=2,
-        worker_mem=200 * 1024**2,
-        supervisor_mem=1 * 1024**3,
-        config={
-            "scheduling.autoscale.enabled": True,
-            "scheduling.autoscale.scheduler_backlog_timeout": 1,
-            "scheduling.autoscale.worker_idle_timeout": 10000000,
-            "scheduling.autoscale.max_workers": 10,
-        },
-    )
-    async with client:
-
-        def time_consuming(x):
-            time.sleep(1)
-            return x * x
-
-        series_size = 100
-        assert (
-            md.Series(list(range(series_size)), chunk_size=1)
-            .apply(time_consuming)
-            .sum()
-            .execute()
-            .fetch()
-            == pd.Series(list(range(series_size))).apply(lambda x: x * x).sum()
-        )
-        autoscaler_ref = mo.create_actor_ref(
-            uid=AutoscalerActor.default_uid(),
-            address=client._cluster.supervisor_address,
-        )
-        assert await autoscaler_ref.get_dynamic_worker_nums() > 0
-
-
-@pytest.mark.timeout(timeout=600)
-@pytest.mark.parametrize(
-    "ray_large_cluster", [{"num_nodes": 2, "num_cpus": 4}], indirect=True
-)
-@require_ray
-@pytest.mark.asyncio
-async def test_auto_scale_in(ray_large_cluster):
-    config = _load_config()
-    config["scheduling"]["autoscale"]["enabled"] = True
-    config["scheduling"]["autoscale"]["worker_idle_timeout"] = 1
-    config["scheduling"]["autoscale"]["max_workers"] = 4
-    config["scheduling"]["autoscale"]["min_workers"] = 2
-    client = await new_cluster(
-        worker_num=0,
-        worker_cpu=2,
-        worker_mem=200 * 102**2,
-        supervisor_mem=1 * 1024**3,
-        config=config,
-    )
-    async with client:
-        autoscaler_ref = mo.create_actor_ref(
-            uid=AutoscalerActor.default_uid(),
-            address=client._cluster.supervisor_address,
-        )
-        new_worker_nums = 3
-        await asyncio.gather(
-            *[autoscaler_ref.request_worker() for _ in range(new_worker_nums)]
-        )
-        series_size = 100
-        assert (
-            md.Series(list(range(series_size)), chunk_size=20).sum().execute().fetch()
-            == pd.Series(list(range(series_size))).sum()
-        )
-        while await autoscaler_ref.get_dynamic_worker_nums() > 2:
-            dynamic_workers = await autoscaler_ref.get_dynamic_workers()
-            print(f"Waiting workers {dynamic_workers} to be released.")
-            await asyncio.sleep(1)
-        await asyncio.sleep(1)
-        assert await autoscaler_ref.get_dynamic_worker_nums() == 2
-
-
-@pytest.mark.skip("Enable it when ray ownership bug is fixed")
-@pytest.mark.timeout(timeout=200)
-@pytest.mark.parametrize("ray_large_cluster", [{"num_nodes": 4}], indirect=True)
-@require_ray
-@pytest.mark.asyncio
-async def test_ownership_when_scale_in(ray_large_cluster):
-    client = await new_cluster(
-        worker_num=0,
-        worker_cpu=2,
-        worker_mem=200 * 1024**2,
-        supervisor_mem=200 * 1024**2,
-        config={
-            "scheduling.autoscale.enabled": True,
-            "scheduling.autoscale.scheduler_check_interval": 1,
-            "scheduling.autoscale.scheduler_backlog_timeout": 1,
-            "scheduling.autoscale.worker_idle_timeout": 10,
-            "scheduling.autoscale.min_workers": 1,
-            "scheduling.autoscale.max_workers": 4,
-        },
-    )
-    async with client:
-        autoscaler_ref = mo.create_actor_ref(
-            uid=AutoscalerActor.default_uid(),
-            address=client._cluster.supervisor_address,
-        )
-        await asyncio.gather(*[autoscaler_ref.request_worker() for _ in range(2)])
-        df = md.DataFrame(mt.random.rand(100, 4, chunk_size=2), columns=list("abcd"))
-        print(df.execute())
-        assert await autoscaler_ref.get_dynamic_worker_nums() > 1
-        while await autoscaler_ref.get_dynamic_worker_nums() > 1:
-            dynamic_workers = await autoscaler_ref.get_dynamic_workers()
-            print(f"Waiting workers {dynamic_workers} to be released.")
-            await asyncio.sleep(1)
-        # Test data on node of released worker can still be fetched
-        pd_df = df.to_pandas()
-        groupby_sum_df = df.rechunk(40).groupby("a").sum()
-        print(groupby_sum_df.execute())
-        while await autoscaler_ref.get_dynamic_worker_nums() > 1:
-            dynamic_workers = await autoscaler_ref.get_dynamic_workers()
-            print(f"Waiting workers {dynamic_workers} to be released.")
-            await asyncio.sleep(1)
-        assert df.to_pandas().to_dict() == pd_df.to_dict()
-        assert (
-            groupby_sum_df.to_pandas().to_dict() == pd_df.groupby("a").sum().to_dict()
-        )
-
-
-@require_ray
-@pytest.mark.asyncio
-def test_init_metrics_on_ray(ray_start_regular, create_cluster):
+def test_init_metrics_on_ray(ray_start_regular_shared, create_cluster):
     client = create_cluster[0]
     assert client.session
     from ....metrics import api
