+++ mars/deploy/oscar/tests/test_ray_dag.py
@@ -120,7 +120,7 @@ async def test_iterative_tiling(ray_start_regular_shared2, create_cluster):
 
 @require_ray
 @pytest.mark.parametrize("config", [{"backend": "ray"}])
-def test_sync_execute(config):
+def test_sync_execute(ray_start_regular_shared2, config):
     test_local.test_sync_execute(config)
 
 
@@ -143,7 +143,7 @@ def test_cancel(ray_start_regular_shared2, create_cluster, test_func):
 
 @require_ray
 @pytest.mark.parametrize("config", [{"backend": "ray"}])
-def test_executor_context_gc(config):
+def test_executor_context_gc(ray_start_regular_shared2, config):
     session = new_session(
         backend=config["backend"],
         n_cpu=2,
